Add ExtensionClass stuff for my tests at an ExtensionClass based pygtk.
Note that these are not the original ExtensionClass files as distributed by
Digital Creations -- they have been modified to allow subclassing of
ExtensionClasses in C as well as in python.  They are binary and source
backward compatible with DC's versions.  Hopefully the patches will be
included in a future official version.